requirements.txt
Flask==3.0.0
supabase==2.3.0
python-dotenv==1.0.0
werkzeug==3.0.1
gotrue==2.1.0
postgrest==0.16.0
storage3==0.7.0